requirements.txt
sounddevice
numpy
spicy
noisereduce
openai-whisper
transformers
torch
safetensors
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib
argon2-cffi
python-multipart
pydantic
pydantic-settings
email-validator
soundfile
sqlalchemy
pymysql